moved sub commands to own classes
