+++ src/main/java/de/kittybot/kittybot/commands/dev/DevCommand.java
@@ -1,25 +1,8 @@
 package de.kittybot.kittybot.commands.dev;
 
-import de.kittybot.kittybot.modules.CommandsModule;
-import de.kittybot.kittybot.objects.enums.Environment;
+import de.kittybot.kittybot.commands.dev.dev.*;
 import de.kittybot.kittybot.slashcommands.application.Category;
 import de.kittybot.kittybot.slashcommands.application.Command;
-import de.kittybot.kittybot.slashcommands.application.CommandOptionChoice;
-import de.kittybot.kittybot.slashcommands.application.options.CommandOptionInteger;
-import de.kittybot.kittybot.slashcommands.application.options.CommandOptionLong;
-import de.kittybot.kittybot.slashcommands.application.options.CommandOptionString;
-import de.kittybot.kittybot.slashcommands.application.options.SubCommand;
-import de.kittybot.kittybot.slashcommands.interaction.GuildInteraction;
-import de.kittybot.kittybot.slashcommands.interaction.Interaction;
-import de.kittybot.kittybot.slashcommands.interaction.Options;
-import de.kittybot.kittybot.slashcommands.interaction.response.FollowupMessage;
-import de.kittybot.kittybot.slashcommands.interaction.response.InteractionResponse;
-import de.kittybot.kittybot.utils.Colors;
-import de.kittybot.kittybot.utils.Config;
-import net.dv8tion.jda.api.EmbedBuilder;
-
-import java.time.Instant;
-import java.util.concurrent.TimeUnit;
 
 @SuppressWarnings("unused")
 public class DevCommand extends Command{
@@ -32,125 +15,8 @@ public class DevCommand extends Command{
 			new RemoveCommand(),
 			new StatsCommand(),
 			new TestCommand(),
-			new PresenceCommand(),
-			new AnnounceCommand()
+			new PresenceCommand()
 		);
-	}
-
-	private static class DeployCommand extends SubCommand{
-
-		public DeployCommand(){
-			super("deploy", "Deploys slash commands to the specified environment");
-			addOptions(
-				new CommandOptionInteger("environment", "In which environment should the commands get deployed").required()
-					.addChoices(
-						new CommandOptionChoice<>("global", 0),
-						new CommandOptionChoice<>("guild", 1)
-					),
-				new CommandOptionLong("guild", "In which guild commands should get deployed")
-			);
-			devOnly();
-		}
-
-		@Override
-		public void run(Options options, Interaction ia){
-			var environment = options.getInt("environment");
-			if(environment == 0){
-				ia.reply(new InteractionResponse.Builder().ephemeral().setContent("processing...").build());
-				ia.getModules().schedule(() -> {
-					ia.get(CommandsModule.class).deployAllCommands(-1L);
-					ia.followup(new FollowupMessage.Builder().setEmbeds(new EmbedBuilder().setColor(Colors.KITTYBOT_BLUE).setDescription("Deployed slash commands globally").build()).build());
-				}, 0, TimeUnit.SECONDS);
-				return;
-			}
-			var guildId = options.has("guild") ? options.getLong("guild") : ia instanceof GuildInteraction ? ((GuildInteraction) ia).getGuildId() : -1L;
-			if(guildId == -1L){
-				ia.error("Please provide a valid guild id");
-				return;
-			}
-			ia.reply(new InteractionResponse.Builder().ephemeral().setContent("processing...").build());
-			ia.getModules().schedule(() -> {
-				ia.get(CommandsModule.class).deployAllCommands(guildId);
-				ia.followup(new FollowupMessage.Builder().setEmbeds(new EmbedBuilder().setColor(Colors.KITTYBOT_BLUE).setDescription("Deployed slash commands for guild `" + guildId + "`").build()).build());
-			}, 0, TimeUnit.SECONDS);
-		}
-
-	}
-
-	private static class RemoveCommand extends SubCommand{
-
-		public RemoveCommand(){
-			super("remove", "Removes slash commands from a specified environment");
-			addOptions(
-				new CommandOptionInteger("environment", "In which environment should the commands get removed").required()
-					.addChoices(
-						new CommandOptionChoice<>("global", 0),
-						new CommandOptionChoice<>("guild", 1)
-					),
-				new CommandOptionLong("guild", "In which guild commands should get removed")
-			);
-			devOnly();
-		}
-
-		@Override
-		public void run(Options options, Interaction ia){
-			var environment = options.getInt("environment");
-			if(environment == 0){
-				if(Environment.getCurrent() == Environment.PRODUCTION){
-					ia.reply(new InteractionResponse.Builder().ephemeral().setContent("Removing commands globally in production is not allowed sorry :3").build());
-					return;
-				}
-				ia.reply(new InteractionResponse.Builder().ephemeral().setContent("processing...").build());
-				ia.getModules().schedule(() -> {
-					var commandsModule = ia.get(CommandsModule.class);
-					commandsModule.deleteAllCommands(-1L);
-					ia.followup(new FollowupMessage.Builder().setEmbeds(new EmbedBuilder().setColor(Colors.KITTYBOT_BLUE).setDescription("Removed slash commands globally").build()).build());
-				}, 0, TimeUnit.SECONDS);
-				return;
-			}
-			var guildId = options.has("guild") ? options.getLong("guild") : ia instanceof GuildInteraction ? ((GuildInteraction) ia).getGuildId() : -1L;
-			if(guildId == -1){
-				ia.error("Please provide a valid guild id");
-				return;
-			}
-			ia.reply(new InteractionResponse.Builder().ephemeral().setContent("processing...").build());
-			ia.getModules().schedule(() -> {
-				var commandsModule = ia.get(CommandsModule.class);
-				commandsModule.deleteAllCommands(guildId);
-				ia.followup(new FollowupMessage.Builder().setEmbeds(new EmbedBuilder().setColor(Colors.KITTYBOT_BLUE).setDescription("Removed slash commands for guild `" + guildId + "`").build()).build());
-			}, 0, TimeUnit.SECONDS);
-		}
-
-	}
-
-	private static class AnnounceCommand extends SubCommand{
-
-		public AnnounceCommand(){
-			super("announce", "Announces a message in all servers");
-			addOptions(
-				new CommandOptionString("message", "The message to announce")
-			);
-			devOnly();
-		}
-
-		@Override
-		public void run(Options options, Interaction ia){
-			var message = options.getString("message");
-			var embed = new EmbedBuilder()
-				.setColor(Colors.KITTYBOT_BLUE)
-				.setAuthor("KittyBot Update", ia.getJDA().getSelfUser().getEffectiveAvatarUrl(), Config.ORIGIN_URL)
-				.setDescription(message)
-				.setTimestamp(Instant.now())
-				.build();
-			ia.getModules().getShardManager().getGuildCache().forEach(guild -> {
-				var channel = guild.getDefaultChannel();
-				if(channel == null){
-					return;
-				}
-				channel.sendMessage(embed).queue();
-			});
-		}
-
 	}
 
 }
