+++ src/main/java/de/kittybot/kittybot/commands/dev/dev/EvalCommand.java
@@ -1,4 +1,4 @@
-package de.kittybot.kittybot.commands.dev;
+package de.kittybot.kittybot.commands.dev.dev;
 
 import de.kittybot.kittybot.slashcommands.application.options.CommandOptionString;
 import de.kittybot.kittybot.slashcommands.application.options.SubCommand;
