+++ src/main/java/de/kittybot/kittybot/commands/dev/dev/StatsCommand.java
@@ -1,4 +1,4 @@
-package de.kittybot.kittybot.commands.dev;
+package de.kittybot.kittybot.commands.dev.dev;
 
 import de.kittybot.kittybot.modules.*;
 import de.kittybot.kittybot.slashcommands.application.options.SubCommand;
